+++ src/components/PlanDetails/index.tsx
@@ -10,11 +10,11 @@ import { MainCard } from '@digitalaidseattle/mui';
 import { Box, Button, Stack, Step, StepLabel, Stepper } from '@mui/material';
 import { createContext, useContext, useState } from "react";
 import { planService } from '../../api/cePlanService';
-import { PlanContext } from '../../pages/plan';
 import { GroupBoard } from "./GroupBoard";
 import { GroupSize } from './GroupSize';
 import { SetupStudents } from './SetupStudents';
 import { TextEdit } from '../TextEdit';
+import { PlanContext } from '../../pages/plan/PlanContext';
 
 // NEW : Add StepperContext to memorize GroupSize
